backend/apps/conversations/api/serializers/conversation.py
from rest_framework import serializers

from apps.conversations.models import Conversation


class ConversationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Conversation
        fields = "__all__"
